+++ tmp.py
@@ -1,0 +1,7 @@
+from typing import Union
+from typing import TypedDict
+
+class PropertiesOptional14769547492369934134(TypedDict, total=False):
+    """"""
+    hello: Union[float, int]
+